feat: add vehicle license selection form with location tracking and area coverage components
